+++ src/SOMcreator/Template/template.txt
@@ -27,7 +27,7 @@ if (isContainer == true && isComposite == false) {
         {%for attribute in attribute_list%}
 
         attributeName= '{{attribute.name}}';
-        returnFormat= '{{attribute.data_type}}';
+        returnFormat= '{{xs_dict[attribute.data_type]}}';
         existingData = pset_dict[propertySetName][attributeName]
         {%if attribute.value_type == constants.LIST%}
         var list = [
